.github/agent/requirements.txt
PyYAML>=6.0
openai>=1.0.0
anthropic>=0.18.0
